changed some copywriting
