+++ apps/web/public/static/locales/en/common.json
@@ -2161,5 +2161,7 @@
   "instant_tab_title": "Instant Booking",
   "instant_event_tab_description": "Let people book immediately",
   "uprade_to_create_instant_bookings": "Upgrade to Enterprise and let guests join an instant call that attendees can jump straight into",
+  "dont_want_to_wait": "Don't want to wait?",
+  "call_immediately": "Call immediately",
   "ADD_NEW_STRINGS_ABOVE_THIS_LINE_TO_PREVENT_MERGE_CONFLICTS": "↑↑↑↑↑↑↑↑↑↑↑↑↑ Add your new strings above here ↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑"
 }
